requirements.txt
bcrypt==4.0.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.2.3
Flask-Ext==0.1
Flask-Login==0.6.2
Flask-LoginManager==1.1.6
greenlet==2.0.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
packaging==23.1
pandas==2.0.0
Pillow==9.4.0
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3
pywin32==305.1
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.6
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.3
wincertstore==0.2
zipp==3.15.0
